+++ rotkehlchen/chain/ethereum/gitcoin/constants.py
@@ -1,1 +1,0 @@
-GITCOIN_GRANTS_PREFIX = 'gitcoingrants'