+++ rotkehlchen/db/schema.py
@@ -417,39 +417,6 @@ CREATE TABLE IF NOT EXISTS ledger_actions (
 );
 """
 
-# Custom enum table for gicoin transaction types
-DB_CREATE_GITCOIN_TX_TYPE = """
-CREATE TABLE IF NOT EXISTS gitcoin_tx_type (
-  type    CHAR(1)       PRIMARY KEY NOT NULL,
-  seq     INTEGER UNIQUE
-);
-/* Ethereum Transaction */
-INSERT OR IGNORE INTO gitcoin_tx_type(type, seq) VALUES ('A', 1);
-/* ZKSync Transaction */
-INSERT OR IGNORE INTO gitcoin_tx_type(type, seq) VALUES ('B', 2);
-"""
-
-
-DB_CREATE_LEDGER_ACTIONS_GITCOIN_DATA = """
-CREATE TABLE IF NOT EXISTS ledger_actions_gitcoin_data (
-    parent_id INTEGER NOT NULL,
-    tx_id TEXT NOT NULL UNIQUE,
-    grant_id INTEGER NOT NULL,
-    clr_round INTEGER,
-    tx_type NOT NULL DEFAULT('A') REFERENCES gitcoin_tx_type(type),
-    FOREIGN KEY(parent_id) REFERENCES ledger_actions(identifier) ON DELETE CASCADE ON UPDATE CASCADE,
-    PRIMARY KEY(parent_id, tx_id, grant_id)
-);
-"""  # noqa: E501
-
-DB_CREATE_GITCOIN_GRANT_METADATA = """
-CREATE TABLE IF NOT EXISTS gitcoin_grant_metadata (
-    grant_id INTEGER NOT NULL PRIMARY KEY,
-    grant_name TEXT NOT NULL,
-    created_on INTEGER NOT NULL
-);
-"""
-
 DB_CREATE_ETHEREUM_TRANSACTIONS = """
 CREATE TABLE IF NOT EXISTS ethereum_transactions (
     tx_hash BLOB NOT NULL PRIMARY KEY,
@@ -860,9 +827,6 @@ BEGIN TRANSACTION;
 {DB_CREATE_ACTION_TYPE}
 {DB_CREATE_IGNORED_ACTIONS}
 {DB_CREATE_BALANCER_EVENTS}
-{DB_CREATE_LEDGER_ACTIONS_GITCOIN_DATA}
-{DB_CREATE_GITCOIN_TX_TYPE}
-{DB_CREATE_GITCOIN_GRANT_METADATA}
 {DB_CREATE_NFTS}
 {DB_CREATE_COMBINED_TRADES_VIEW}
 COMMIT;